Move Tag* classes to new metadatabox/tagdiff.py
